feat(back): organize DTOs into subfolders
- Group DTOs by model: user, topic, post, comment
