+++ back/src/main/java/com/openclassrooms/mddapi/dto/comment/CommentResponseDTO.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.mddapi.dto;
+package com.openclassrooms.mddapi.dto.comment;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -6,28 +6,19 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 @RequiredArgsConstructor
-public class PostResponseDTO {
+public class CommentResponseDTO {
 
     final private Long id;
 
-    @JsonProperty("topic_id")
-    final private Long topicId;
-
-    @JsonProperty("user_id")
-    final private Long userId;
-
-    final private String title;
+    final private String username;
 
     final private String content;
 
     @JsonProperty("created_at")
     @JsonFormat(pattern = "yyyy/MM/dd")
     final private LocalDateTime createdAt;
-
-    final private List<CommentResponseDTO> comments;
 
 }
