+++ back/src/main/java/com/openclassrooms/mddapi/dto/topic/TopicDTO.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.mddapi.dto;
+package com.openclassrooms.mddapi.dto.topic;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
